SQL Base Framework
This is the SQL file that creates all the tables in the schema with necessary constraints. Unfortunately, I was struggling to find a way to input the data seamlessly without just typing hundreds of lines out. This should be a good place for you guys to start on your parts.